fix: Ensure refresh after unexpected not-found works in all cases Motivation
An issue was reported in production whereby models would stop working following restarts of the model runtime container, despite "recovery" logic intended to insulate against this. It turns out that this was because the runtime (model server) in question was returning NOT_FOUND from calls unloadModel when the model wasn't loaded, rather than completing successfully. This should not be a problem because model-mesh was meant to treat the two responses as equivalent in this case, however due to an ordering mistake some post-unload logic wasn't being invoked.

The auto-recovery after missing model behaviour would still actually work the first time for a given model in a given pod, but if the collocated runtime restarted a second time then it could hit this failure state.

Further, the buggy code path was only taken when the internal serving and management interfaces were different, which they were in this case (UDS being used for serving only).

Modifications

- Reorder the Future processing chain in SidecarModelMesh for UnloadModelResponses so that NOT_FOUND errors are handled and converted to "success" responses prior to other processing.
- Add/extend unit tests to cover "refresh missing model" behaviour, with permutations of runtime configuration/behaviour including the one that triggered this bug.

Result

Model-mesh is resilient as intended in this particular configuration.