feat: Add Kube NetworkPolicy to sample deployment configs (#527)
Motivation

Kubernetes NetworkPolicys are being increasingly used to lock down intra-cluster network communications. The model-mesh network flows are quite simple and easy to capture in such a policy, so it would be good to extend our example deployment configs to include this.

Modifications

Add a NetworkPolicy to the kustomize-based example deployment configuration, update the README.md in that dir with some info about it.

Result

A NetworkPolicy yaml definition that consuming teams can refer to.