fix: Fix cascading eviction edge case with large model size variations  (#530)
* fix: Fix cascading eviction edge case with large model size variations

Motivation

A large consumer of model-mesh have been increasingly encountering cascading evictions, where all of the models in a particular pod will be unnecessarily flushed at once.

This is something which had been encountered/fixed in the past, but debugging and closer inspection highlighted another place in the code that requires similar handling. Specifically if the post-loading size of a model is much larger than the pre-load predicted size, its cache entry will be inflated to the new size without any cascading-eviction protection.

There is an reserved unloading buffer which mitigates this in most cases, but less so when the size of mis-prediction is large.

Modifications

- Move all of the logic related to managing the unloading buffer to a separate ModelCacheUnloadBufManager class
- Introduce a "cache deficit" value to track when we have over-allocations of the kind described above. "borrow" this extra space from the unloading buffer cache entry and ensure that it's the first thing to be repaid when space is freed (i.e. usually on completion of model unloads)
- Add a unit test to exercise this case

Result

Fixes https://github.ibm.com/ai-foundation/wml-serving/issues/1173