fix: Fix incorrect values produced for some metrics
Motivation
We discovered that the values reported for the loading, unloading and cache-miss-delay prometheus metrics were very wrong. This was due to the recorded time durations being incorrectly treated as absolute start times.

Modifications
Change calls to the logTimingMetricSince method in three places to use logTimingMetricDuration instead.

Result
Correct times are now recorded for these metrics.

Signed-off-by: Angel Luu <angel.luu@us.ibm.com>