fix: Strengthen ConcurrentLinkedHashMap write semantics
Motivation

Model-mesh uses a modified version of the open source ConcurrentLinkedHashMap library for its LRU cache in each node. This is a concurrent data structure designed for caching, which minimizes locking done on each cache read/write. It is achieved by maintaining a buffer/queue of updates which are processed in batches rather than locking the ordering structure on every call. However, this optimization means that the effects of a given cache update may not be immediately visible - in particular the weighted size changes resulting from entry additions/adjustments and removals.

This looks to be the cause of a race condition which can trigger cascading evictions, since the bookkeeping/synchronization that modelmesh layers on top of the cache assumes stronger semantics for writes.

For the model-mesh use case, write contention is not a concern since they are infrequent relative to reads (loading/unloading a new model vs inferencing requests against an existing model).

Modifications

Simplify the internal implementation of the modified ConcurrentLinkedHashMap such that writes take effect immediately (i.e. by the thread calling the corresponding method). This involves removing the writeBuffer and associated logic, and taking the global lock on each write.

The writeBuffer was actually also used for reads that included a specific timestamp since the readBuffers only hold the Node to be updated. To accommodate these, the readBuffer arrays were generalized to hold an object that could be either a Node or a ReadRecord, the latter a new class just to hold a Node and associated read timestamp.

Result

Hopefully no more race condition that causes cascading evictions.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>