fix: Fix unintentional forwarding to failing instances
Motivation

A bug was discovered where inference requests are incorrectly forwarded to failing instances in the case that (a) there are instances in the failing list (b) the instance doing the routing is the one that _should_ process the request.

This was due to how the custom litelinks loadbalancers used by model-mesh return null to indicate that the RPC should be aborted and instead handled by the local instance. In these cases litelinks will first attempt sending to failing instances if there are any.

Modifications

Take advantage of new ABORT_REQUEST functionality added to litelinks 1.7.2. The load balancer logic now returns ABORT_REQUEST rather than null when the local instance has been chosen to handle the request.

Also make some local variables final within the LB implementations to aid with readability.

Result

Less disruptive behaviour in some failure modes.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>