fix: Hopefully final fix for cascading eviction
Motivation

We had hoped that the notorious cascading eviction bug was fixed recently but though still infrequent it was observed once again in the wild.

Upon close analysis of these latest occurrences in production logs it was possible to determine the source of the remaining race condition.

The problem basically comes down to the fact that we attempt to "immediately" adjust the unload buffer accounting at the point entries are evicted from the cache, but this was still not done atomically. So with the right confluence of events between different threads, it was possible for a model loading thread to incorrectly think there's sufficient space to proceed rather than waiting for other unloads to complete. This in turn results in the over-committed state that causes the eviction cascade.

Modifications

- Unify the cache's eviction lock with the separate UNLOAD_BUF lock, and move the eviction callbacks within the locked context.
- Eliminate the separate eviction thread and instead perform the eviction processing inline (including related cache accounting updates).
- Add unit tests to exercise eviction / unload buffer behaviour and to trigger an eviction cascade prior to this fix.

Result

Hopefully no more cascading evictions possible.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>