fix: Timeouts on all internal requests; better timeout logging
Motivation

Some internal requests to trigger model loads are done with no timeout. This had never been an issue because the remote processing is async in nature but it's best to have a safeguard timeout just in case these calls hang for some reason (as was observed in relation to a separate now-fixed deadlock bug).

Modifications

- Add 3 second timeout for the internal ensureLoaded calls and ensure that ensureLoadedInternal() is used consistently for these
- Improve exception/logging messages related to timeouts (in particular avoid misleading "Unexpected internalOperation" messages)
- Add a new "in-" thread name prefix to distinguish from "ex-" where the request originates internally

Result

Internal remote request threads can't hang indefinitely.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>