fix: Avoid possibility of distributed deadlock
Motivation

A deadlock-type state was observed where all of the model loading threads were stuck waiting for space to be freed by unloads, and all of the 10 "task" threads that are used for scheduling unloads were stuck waiting for internal ensureLoaded calls to complete.

It's not clear (yet) why the remote internal ensureLoaded calls were stuck because they do not block wait for the loading to complete. But this was in a test cluster under heavy inferencing and model registration load.

Modifications

- Add 3 second timeout for the internal ensureLoaded calls and ensure that ensureLoadedInternal() is used consistently for these
- Add a separate threadpool for async internal requests to limit how many run in parallel and to "protect" the main taskPool in case they get stuck (even though the timeout is now in place)
- Improve exception/logging messages related to timeouts (in particular avoid misleading "Unexpected internalOperation" messages)
- Add a new "in-" thread name prefix to distinguish from "ex-" where the request originates internally

Result

Avoid possible deadlock situation under heavy load

Signed-off-by: Nick Hill <nickhill@us.ibm.com>