fix: Repair invalid ModelRecord lastUsed values in registry
Motivation

Some ModelRecords were observed in the registry with an invalid lastUsed time of Long.MAX_VALUE, effectively pinning them at the front of the cache and causing them to remain loaded even if not used.

It's not clear whether the bug causing this has yet been fixed - for now we'll repair the entries and can monitor logs for the warning message below.

Modification

Add a check in the reaper loop over all models in the registry, and update invalid records with a lastUsed time of "3 hours ago".

Result

Avoid "stuck" models due to incorrect lastUsed time.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>