fix: Repair invalid ModelRecord lastUsed values in registry
Motivation

Some model records were observed in the registry with an invalid lastUsed time of Long.MAX_VALUE, effectively pinning them at the front of the cache and causing them to remain loaded even if not used.

An attempt to fix this was made in https://github.com/kserve/modelmesh/pull/36, but this didn't end up repairing things in the affected environments since the invalid lastUsed value was still in the local cache(s) and ended up being synchronized back to the registry.

Modifications

- Add some logic to the janitor processing to detect and update cache entries with a Long.MAX_VALUE lastUsed value
   - This requires a new forceSetLastUsedTime() method in ConcurrentLinkedHashMap since it's otherwise not possible to decrease the current value
- Never sync this value back to the registry
- Avoid creating failure placeholder records with a Long.MAX_VALUE lastUsed value

Result

Avoid "stuck" models due to incorrect lastUsed time. We still haven't determined the root cause (which might already be fixed) but will be able to see whether the issue arises again once this fix is deployed.

Signed-off-by: Nick Hill <nickhill@us.ibm.com>