docs: Indicate which model-runtime proto SPI rpcs are optional
Motivation

In the model server gRPC service provider interface, implementation of the predictModelSize RPC is optional, and modelSize doesn't need to be implemented if sizes are always returned in the loadModel response.

This is documented in the modelmesh docs but not in the comments of the service .proto file.

Modifications

Add appropriate wording to doc comments in the model-runtime.proto file to convey the fact these rpcs aren't required.

Result

Clearer docs

Signed-off-by: Nick Hill <nickhill@us.ibm.com>