fix: Ensure intra-cluster InternalExceptions are propagated properly
Motivation

In certain error edge cases, InternalExceptions can be thrown on the inferencing path, but these currently don't propagate properly because they aren't declared to be thrown by the internal thrift applyModel(Multi) method. Because of this, they are treated as generic errors meaning that the error detail (message) is lost and when returned to the user can have a confusing message like "Nowhere available to load".

Modifications

- Add InternalException to internal applyModel thrift rpc method definitions
- Also update to use latest version of thrift

Result

InternalExceptions will propagate properly between modelmesh pods

Signed-off-by: Nick Hill <nickhill@us.ibm.com>