feat: Add vModelId to PayloadProcessor Payload
Motivation

Currently the payloads passed to PayloadProcessors only contain the modelId, which in the case of vModels will be a "resolved" modelId corresponding to a particular model revision (in particular this will be true when used in KServe modelmesh-serving). It would be useful to include the vModelId too.

Modifications

Add a vModelId field to the Payload class and correspondingly update built-in PayloadProcessor implementations where applicable.

It may be null if the request was directed at a concrete modelId rather than a vModelId.

Result

Both modelId and vModelId are available to PayloadProcessors

Signed-off-by: Nick Hill <nickhill@us.ibm.com>